+++ apps/be/src/app.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { PrismaService } from './prisma.service';
 import { HealthModule } from './health/health.module';
-import { UsersController } from './user.controller';
-import { UserService } from './service/user.service';
+import { PrismaModule } from './prisma/prisma.module';
+import { UserController } from './user/user.controller';
+import { UserModule } from './user/user.module';
 
 @Module({
-  imports: [HealthModule],
-  controllers: [AppController, UsersController],
-  providers: [AppService, UserService, PrismaService],
+  imports: [HealthModule, PrismaModule, UserModule],
+  controllers: [AppController],
+  providers: [AppService],
 })
 export class AppModule {}
