+++ apps/be/src/app.service.ts
@@ -1,16 +1,16 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
 import { Employee } from '@prisma/client';
+import { PrismaService } from './prisma/prisma.service';
 
 @Injectable()
 export class AppService {
-  constructor(private prisma: PrismaService) { }
+  constructor(private prisma: PrismaService) {}
 
   getHello(): string {
     return 'Hello World!';
   }
 
   async employee(): Promise<Employee[]> {
-    return this.prisma.employee.findMany()
+    return this.prisma.employee.findMany();
   }
 }
