+++ apps/be/src/user/user.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { UserService } from './user.service';
+import { UserController } from './user.controller';
+import { PrismaModule } from 'src/prisma/prisma.module';
+
+@Module({
+  imports: [PrismaModule],
+  controllers: [UserController],
+  providers: [UserService, UserController],
+})
+export class UserModule {}
