+++ apps/be/src/user/user.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import User from 'src/domain/user';
+import User from 'src/user/domain/user';
 
 @Injectable()
 export class UserService {
